bot/requirements.txt
python_telegram_bot==20.1
python-dotenv
requests
pandas
openpyxl
howlongtobeatpy
rawg
gspread
mysql-connector-python
pymysql
SQLAlchemy
